+++ src/dftbp/dftbplus/main.F90
@@ -16,7 +16,7 @@ module dftbp_dftbplus_main
   use dftbp_common_globalenv, only : stdOut, withMpi
   use dftbp_common_hamiltoniantypes, only : hamiltonianTypes
   use dftbp_common_status, only : TStatus
-  use dftbp_derivs_numderivs2, only : TNumderivs, next, getHessianMatrix, dipoleAdd
+  use dftbp_derivs_numderivs2, only : TNumderivs, next, getHessianMatrix, dipoleAdd, polAdd
   use dftbp_derivs_perturb, only : TResponse
   use dftbp_dftb_blockpothelper, only : appendBlockReduced
   use dftbp_dftb_boundarycond, only : TBoundaryConditions
@@ -69,11 +69,13 @@ module dftbp_dftbplus_main
       & writeEigenVectors, writeProjectedEigenvectors, writeCurrentGeometry, writeDetailedOut4,&
       & writeEsp, printGeostepInfo, writeDetailedOut2dets, printEnergies, printVolume,&
       & printPressureAndFreeEnergy, writeDetailedOut6, writeDetailedOut7,&
-      & writeFinalDriverstatus, writeHessianout, writeBornChargesOut, writeAutotestTag,&
-      & writeResultsTag, writeDetailedXml, writeCosmoFile, printForceNorm, printLatticeForceNorm,&
-      & writeDerivBandOut, writeDetailedOut8, writeDetailedOut9, writeDetailedOut10
+      & writeFinalDriverstatus, writeHessianout, writeBornChargesOut, writeBornDerivs,&
+      & writeAutotestTag, writeResultsTag, writeDetailedXml, writeCosmoFile, printForceNorm,&
+      & printLatticeForceNorm, writeDerivBandOut, writeDetailedOut8, writeDetailedOut9,&
+      & writeDetailedOut10
   use dftbp_dftbplus_outputfiles, only : autotestTag, bandOut, fCharges, fShifts, fStopScc, mdOut,&
-      & userOut, fStopDriver, hessianOut, bornChargesOut, resultsTag, derivEBandOut
+      & userOut, fStopDriver, hessianOut, bornChargesOut, bornDerivativesOut, resultsTag,&
+      & derivEBandOut
   use dftbp_dftbplus_qdepextpotproxy, only : TQDepExtPotProxy
   use dftbp_dftbplus_transportio, only : readShifts, writeShifts, writeContactShifts
   use dftbp_elecsolvers_elecsolvers, only : TElectronicSolver
@@ -189,7 +191,7 @@ contains
     logical :: isUnReduced
 
     type(TStatus) :: errStatus
-    real(dp), pointer :: pDynMatrix(:,:), pDipDerivMatrix(:,:)
+    real(dp), pointer :: pDynMatrix(:,:), pDipDerivMatrix(:,:), pPolDerivMatrix(:,:,:)
 
     call initGeoOptParameters(this%tCoordOpt, this%nGeoSteps, tGeomEnd, tCoordStep, tStopDriver,&
         & iGeoStep, iLatGeoStep)
@@ -285,10 +287,31 @@ contains
             & this%qiBlockIn, this%deltaRhoIn, size(this%iAtInCentralRegion), this%multipoleInp)
       end if
 
-      if (this%tForces) then
-        if (this%tDipole.and.allocated(this%derivDriver)) then
-          call dipoleAdd(this%derivDriver, this%dipoleMoment)
+      if (this%tDipole.and.allocated(this%derivDriver)) then
+        call dipoleAdd(this%derivDriver, this%dipoleMoment)
+      end if
+
+      if (this%doPerturbEachGeom.and.allocated(this%derivDriver)) then
+
+        if (this%isEResp) then
+          call this%response%wrtEField(env, this%parallelKS, this%filling, this%eigen,&
+              & this%eigVecsReal, this%eigvecsCplx, this%ints%hamiltonian, this%ints%overlap,&
+              & this%orb, this%nAtom, this%species, this%neighbourList, this%nNeighbourSK,&
+              & this%denseDesc, this%iSparseStart, this%img2CentCell, this%coord, this%scc,&
+              & this%maxSccIter, this%sccTol, this%isSccConvRequired, this%nMixElements,&
+              & this%nIneqOrb, this%iEqOrbitals, this%tempElec, this%Ef, this%spinW,&
+              & this%thirdOrd, this%dftbU, this%iEqBlockDftbu, this%onSiteElements,&
+              & this%iEqBlockOnSite, this%rangeSep, this%nNeighbourLC, this%pChrgMixer,&
+              & this%kPoint, this%kWeight, this%iCellVec, this%cellVec, this%polarisability,&
+              & this%dEidE, this%dqOut, this%neFermi, this%dEfdE, errStatus, this%dynRespEFreq)
+          if (errStatus%hasError()) then
+            call error(errStatus%message)
+          end if
+          call polAdd(this%derivDriver, this%polarisability(:,:,1))
         end if
+      end if
+
+      if (this%tForces) then
         call getNextGeometry(this, env, iGeoStep, tWriteRestart, constrLatDerivs, tCoordStep,&
             & tGeomEnd, tStopDriver, iLatGeoStep, tempIon, tExitGeoOpt)
         if (tExitGeoOpt) then
@@ -347,9 +370,17 @@ contains
 
     if (env%tGlobalLead .and. this%tDerivs) then
       if (this%tDipole) then
-        call getHessianMatrix(this%derivDriver, pDynMatrix, pDipDerivMatrix)
+        if (this%doPerturbEachGeom) then
+          call getHessianMatrix(this%derivDriver, pDynMatrix, pDipDerivMatrix, pPolDerivMatrix)
+        else
+          call getHessianMatrix(this%derivDriver, pDynMatrix, pDipDerivMatrix)
+        end if
       else
-        call getHessianMatrix(this%derivDriver, pDynMatrix)
+        if (this%doPerturbEachGeom) then
+          call getHessianMatrix(this%derivDriver, pDynMatrix, pol=pPolDerivMatrix)
+        else
+          call getHessianMatrix(this%derivDriver, pDynMatrix)
+        end if
       end if
       call writeHessianOut(hessianOut, pDynMatrix, this%indMovedAtom, errStatus)
       if (errStatus%hasError()) then
@@ -362,15 +393,25 @@ contains
         if (errStatus%hasError()) then
           call error(errStatus%message)
         end if
-        if (env%tGlobalLead .and. this%tWriteDetailedOut) then
+        if (this%tWriteDetailedOut) then
           call writeDetailedOut8(this%fdDetailedOut%unit,&
               & this%eFieldScaling%scaledSoluteDipole(pDipDerivMatrix))
+        end if
+      end if
+      if (this%doPerturbEachGeom) then
+        call writeBornDerivs(bornDerivativesOut, pPolDerivMatrix, this%indMovedAtom,&
+            & size(this%iAtInCentralRegion), errStatus)
+        if (errStatus%hasError()) then
+          call error(errStatus%message)
         end if
       end if
     else
       nullify(pDynMatrix)
       if (this%tDipole) then
         nullify(pDipDerivMatrix)
+      end if
+      if (this%doPerturbEachGeom) then
+        nullify(pPolDerivMatrix)
       end if
     end if
 
@@ -425,7 +466,7 @@ contains
 
   #:endif
 
-    if (this%isDFTBPT) then
+    if (this%doPerturbation) then
 
       if (this%isEResp) then
         call this%response%wrtEField(env, this%parallelKS, this%filling, this%eigen,&